# docs/swagger.yaml
definitions:
  entity.AIAnalyticsRequest:
    properties:
      deep_work:
        $ref: '#/definitions/entity.DeepWorkData'
      domains:
        items:
          type: string
        minItems: 1
        type: array
      domains_count:
        minimum: 1
        type: integer
      engagement_rate:
        maximum: 100
        minimum: 0
        type: number
      period:
        type: string
      timestamp:
        type: string
      tracked_hours:
        minimum: 0
        type: number
      user_id:
        type: string
    required:
    - deep_work
    - domains
    - domains_count
    - engagement_rate
    - tracked_hours
    type: object
  entity.BatchCreateUserBehaviorRequest:
    properties:
      events:
        items:
          $ref: '#/definitions/entity.CreateUserBehaviorRequest'
        type: array
    required:
    - events
    type: object
...
  entity.CreateUserBehaviorRequest:
    properties:
      key:
        type: string
      sessionId:
        type: string
      ts:
        type: string
      type:
        type: string
      url:
        type: string
      userId:
        type: string
      userName:
        type: string
      x:
        type: integer
      "y":
        type: integer
    required:
    - sessionId
    - ts
    - type
    - url
    type: object
...
  entity.DomainAnalysis:
    properties:
      analysis:
        $ref: '#/definitions/entity.DetailedAnalysis'
      focus_insight:
        type: string
      focus_level:
        type: string
      recommendations:
        items:
          type: string
        type: array
      work_pattern:
        type: string
    type: object
...
  entity.FocusLevelResponse:
    properties:
      focus_level:
        type: string
      insight:
        type: string
      method:
        description: '"ai" или "fallback"'
        type: string
      timestamp:
        type: string
    type: object
  entity.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/entity.PaginationInfo'
      success:
        type: boolean
    type: object
  entity.PaginationInfo:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  entity.UserBehavior:
    properties:
      createdAt:
        type: string
      id:
        type: string
      key:
        type: string
      sessionId:
        type: string
      ts:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      url:
        type: string
      userId:
        type: string
      userName:
        type: string
      x:
        type: integer
      "y":
        type: integer
    required:
    - sessionId
    - ts
    - type
    - url
    type: object
...
  wrapper.ErrorWrapper:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  wrapper.ResponseWrapper:
    properties:
      data: {}
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
  /ai-analytics/domain-usage:
    post:
      consumes:
      - application/json
      description: Get AI-powered analysis of user's domain usage patterns, productivity
        insights, and recommendations
      parameters:
      - description: Analytics request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.AIAnalyticsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.DomainAnalysis'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Analyze domain usage with AI
      tags:
      - /api/v1/admin/ai-analytics
  /ai-analytics/focus-level:
    get:
      consumes:
      - application/json
      description: Get basic focus level assessment based on domain count (fallback
        method)
      parameters:
      - description: Number of unique domains
        in: query
        name: domains_count
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.FocusLevelResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Get focus level without AI
      tags:
      - /api/v1/admin/ai-analytics
  /behaviors:
    get:
      consumes:
      - application/json
      description: Get user behavior events with optional filters
      parameters:
      - description: User ID
        in: query
        name: userId
        type: string
      - description: Session ID
        in: query
        name: sessionId
        type: string
      - description: Event type
        in: query
        name: eventType
        type: string
      - description: URL (partial match)
        in: query
        name: url
        type: string
      - description: Start time (RFC3339 format)
        in: query
        name: startTime
        type: string
      - description: End time (RFC3339 format)
        in: query
        name: endTime
        type: string
      - description: 'Time period filter: ''today'', ''week'', ''month'', ''year'''
        in: query
        name: period
        type: string
      - description: Page number (starts from 1)
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 1000)'
        in: query
        name: per_page
        type: integer
      - description: Limit (deprecated, use per_page)
        in: query
        name: limit
        type: integer
      - description: Offset (deprecated, use page)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entity.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.UserBehavior'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Get user behaviors
      tags:
      - /api/v1/admin/behaviors
    post:
      consumes:
      - application/json
      description: Create a single user behavior event
      parameters:
      - description: Behavior data
        in: body
        name: behavior
        required: true
        schema:
          $ref: '#/definitions/entity.CreateUserBehaviorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.UserBehavior'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Create user behavior event
      tags:
      - /api/v1/inayla/behaviors
  /behaviors/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific user behavior event
      parameters:
      - description: Behavior ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Delete behavior
      tags:
      - /api/v1/admin/behaviors
    get:
      consumes:
      - application/json
      description: Get a specific user behavior event by ID
      parameters:
      - description: Behavior ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.UserBehavior'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Get behavior by ID
      tags:
      - /api/v1/admin/behaviors
  /behaviors/batch:
    post:
      consumes:
      - application/json
      description: Create multiple user behavior events in one request
      parameters:
      - description: Behaviors data
        in: body
        name: behaviors
        required: true
        schema:
          $ref: '#/definitions/entity.BatchCreateUserBehaviorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Batch create user behavior events
      tags:
      - /api/v1/inayla/behaviors
...
  /behaviors/sessions/{sessionId}:
    get:
      consumes:
      - application/json
      description: Get summary information about a specific session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.SessionSummary'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Get session summary
      tags:
      - /api/v1/admin/behaviors
  /behaviors/stats:
    get:
      consumes:
      - application/json
      description: Get statistics about user behaviors
      parameters:
      - description: User ID
        in: query
        name: userId
        type: string
      - description: Session ID
        in: query
        name: sessionId
        type: string
      - description: Event type
        in: query
        name: eventType
        type: string
      - description: URL (partial match)
        in: query
        name: url
        type: string
      - description: Start time (RFC3339 format)
        in: query
        name: startTime
        type: string
      - description: End time (RFC3339 format)
        in: query
        name: endTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.UserBehaviorStats'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Get behavior statistics
      tags:
      - /api/v1/admin/behaviors
  /behaviors/user-events:
    get:
      consumes:
      - application/json
      description: Get statistics about user behaviors
      parameters:
      - description: User ID
        in: query
        name: userId
        type: string
      - description: Start time (RFC3339 format)
        in: query
        name: startTime
        type: string
      - description: End time (RFC3339 format)
        in: query
        name: endTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.UserEventsCountResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Get behavior statistics
      tags:
      - /api/v1/admin/behaviors
  /behaviors/users/{userId}/sessions:
    get:
      consumes:
      - application/json
      description: Get all sessions for a specific user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 50, max: 200)'
        in: query
        name: per_page
        type: integer
      - description: Limit (deprecated, use per_page)
        in: query
        name: limit
        type: integer
      - description: Offset (deprecated, use page)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.SessionSummary'
                  type: array
                pagination:
                  $ref: '#/definitions/entity.PaginationInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Get user sessions
      tags:
      - /api/v1/admin/behaviors
...
  /extension/users:
    get:
      consumes:
      - application/json
      description: Get list of extension users with optional filters
      parameters:
      - description: Filter by username
        in: query
        name: username
        type: string
      - description: Filter by active status
        in: query
        name: isActive
        type: boolean
      - description: Page number (starts from 1)
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 200)'
        in: query
        name: per_page
        type: integer
      - description: Limit (deprecated, use per_page)
        in: query
        name: limit
        type: integer
      - description: Offset (deprecated, use page)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.PaginatedResponseWrapper'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.ExtensionUserPublic'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Get all extension users
      tags:
      - /api/v1/admin/extension
  /extension/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an extension user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Delete extension user
      tags:
      - /api/v1/admin/extension
    get:
      consumes:
      - application/json
      description: Get a specific extension user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.ExtensionUserPublic'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Get extension user by ID
      tags:
      - /api/v1/admin/extension
    put:
      consumes:
      - application/json
      description: Update extension user information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateExtensionUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.ExtensionUserPublic'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Update extension user
      tags:
      - /api/v1/admin/extension
  /extension/users/{id}/regenerate-key:
    post:
      consumes:
      - application/json
      description: Regenerate API key for extension user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.RegenerateAPIKeyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Regenerate API key
      tags:
      - /api/v1/admin/extension
  /extension/users/auth:
    post:
      consumes:
      - application/json
      description: Validate extension user API key
      parameters:
      - description: API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.ExtensionUserPublic'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Validate API key
      tags:
      - /api/v1/inayla/extension
  /extension/users/generate:
    post:
      consumes:
      - application/json
      description: Create a new extension user with API key
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.CreateExtensionUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.ExtensionUser'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Create extension user
      tags:
      - /api/v1/admin/extension
  /extension/users/stats:
    get:
      consumes:
      - application/json
      description: Get statistics about extension users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.ExtensionUserStats'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Get extension users statistics
      tags:
      - /api/v1/admin/extension
...
  /users/auth:
    post:
      consumes:
      - application/json
      description: Create a new user or authenticate an existing user with password
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserWithPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/response.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Create or authenticate user with password
      tags:
      - users
...
  /users/profile:
    get:
      consumes:
      - application/json
      description: Get user by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/response.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Get user by ID
      tags:
      - users
swagger: "2.0"